requirements.txt
diffusers==0.21.1
kornia==0.7.0
matplotlib==3.7.3
numpy==1.24.4
Pillow==10.0.0
PyWavelets==1.4.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.1
